+++ test/faker/crypto_coin_test.exs
@@ -1,0 +1,5 @@
+defmodule Faker.CryptoCoinTest do
+  use ExUnit.Case, async: true
+
+  doctest Faker.CryptoCoin
+end