+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 * Added: Category `Food` for `Armenian Language` [[@hovikman][]]
 * Added: Support for `Armenian Language` [[@hovikman][]]
 * Added: `Faker.Vehicle` [[@daytonn][]]
+* Added: `Faker.CryptoCoin` [[@thadeubrito][]]
 
 ### Changed
 
@@ -316,6 +317,7 @@ Change log itself follows [Keep a CHANGELOG](http://keepachangelog.com) format.
 [@sotojuan]: https://github.com/sotojuan
 [@stfnsr]: https://github.com/stfnsr
 [@tbash]: https://github.com/tbash
+[@thadeubrito]: https://github.com/thadeubrito
 [@theabrad]: https://github.com/theabrad
 [@tobyhinloopen]: https://github.com/tobyhinloopen
 [@vbrazo]: https://github.com/vbrazo